requirements.txt
aiogram~=3.22.0
python-dotenv~=1.1.1
openai
ruff
colorlog